Few things
Staff Tools

Currently only 1 tool what is an profile overlay, goes in the top left corner of your game and shows the players profile, and has a button to show PO's what can easily be accessed with F3 in-game.

Currently can only be accessed if you are staff and you have enabled it in the overlay, and soon as you copy a steamid it will display the profile in top left.

Proper branch handling

All gmod branches (Beta Versions) should now work with the overlays.